refactor(server): Split protocol/messages.rs into focused modules
Split the 1,749-line messages.rs file into focused modules under 500 lines each:

- types.rs (157 lines): Core type definitions (errors, configs, field descriptions)
- backend.rs (309 lines): BackendMessage enum and encode() method
- frontend.rs (265 lines): FrontendMessage enum and decode() methods
- helpers.rs (45 lines): Encoding/decoding utility functions
- mod.rs (26 lines): Module re-exports for API compatibility

Tests organized in dedicated directory:
- tests/encoding.rs (317 lines): Basic encoding/decoding tests
- tests/malformed.rs (424 lines): Security-relevant malformed message tests
- tests/selective_config.rs (270 lines): SelectiveUpdatesConfig parsing tests

No functional changes - pure refactor with all 400 server tests passing.

Closes #4339
