+++ crates/vibesql-server/src/protocol/tests/mod.rs
@@ -1,0 +1,5 @@
+//! Protocol message tests
+
+mod encoding;
+mod malformed;
+mod selective_config;
